SLVS-1816 Move server event files from Core into Connected Mode assembly.
This fixes the architecture violation of Core referencing the SonarQube.Client assembly
